fix: Remove session storage for categories, rename and delete relations
The main bottleneck is not the categories, so no need for session storage,
aslo removed an unnecessary hasMany relation with categories. Now, the
number of queries for categories is low.

The problem is with the autorization relations in pivot table, because
we need to query every one of them, so there are too many queries.
TODO: performance optimalization needed to have less queries!