fix: Replace Gates to Policies for upload/download authorization
All the gates were replaced with policies belonging to the Category
and the User models.

For authorization, the permissions column in users table was used,
because previously a pivot table was containing all of the permis-
sions between users and categories. All permission check made 1
sql request to fetch the permission if it is exists.
This lead to unnecessary sql queries. basically every time there
was a @can permission check, the was a sql query as well. So, if
there are N categories -> there are also N sql queries, plus the 1
query for getting the user, and thus there were N+1 queries every
time the page has reloaded.

Now, the performance is much better without the N+1 problem.